Created executeQuery in Database
- Removed old loadInitialData method from Database
- ExecuteQuery method exists in Database now. To use it, you enter the
name of the query (used for system.out.println), the sql of the query,
and the QueryResultFormat for what you want back.
- Created 3 QueryResultFormats - posResFormat, userResFormat, and
sopResFormat. Each returns an ArrayList of their respective class
(Position, User, and SOP, respectively)
- Those 3 should be used when you want just a single instance of the
classes as well, just perform a check that you only received an
ArrayList of size 1