Created SearchForSOPs Method in Database
- It is designed to handle a bunch of different SOP search queries in
the database. It's designed so that we don't have to create new methods
for each different combination of SOP search criteria given. (e.g. it
covers find SOP by title and priority, which we didn't have a method for
yet)
- I changed the existing SOP search queries to use this new method, and
I made sure all the JUnits still pass for it
- Currently, I've deprecated all those other methods, with the intention
to eventually remove them (for now, I've kept them so as not to break
any work Ryan may have done)
- I plan to do the same for User and Position in the near future as
well.